POST with invalid params should stub :update instead of :save
I noticed a inconsistency in the generated controller scaffold.

The spec for valid params expects a call to `update`, but the spec for invalid params expects a call to `save`. Only 'update' is used with the controller action.

I assume this only worked because the Rails implementation of `update` calls `save` within the implementation at some point?