Fix formatting for feature description.
RelishApp and Cucumber are very picky about syntax. **ANY** line which
starts with an octothorpe is treated as a comment. This makes the lexer
very unhappy.

In order to trick it I had to add a unicode space (specifically the zero
width space <200b>). This alone didn't fix things. For some reason,
using a semi-colon fixed things. I tried several additional variations,
all of which worked; but they each still required the semi-colon:

- `controller { ; }`
- `controller do; end`

I settled on this various as I felt it read a bit better with the
comment being in the `controller` definition.