Include smoke task for non-ActiveRecord apps.
Creates a 2nd set of generator and smoke tasks which target an example
app which does not use ActiveRecord. In order to accomplish this, a
custom in-memory model is used which supports enough of the AR protocol
to allow default specs to pass.

Use a custom "in-memory" model as the ORM replacement. This is a simple
stand-in allowing us to continue to use all of the other Rails
generators. It also prevents us from needing to pull in another ORM gem
such as such as Mongoid, DataMapper, ROM, etc.

The custom "model" is necessary as the non-ActiveRecord example app's
specs fail with `NameError`s. This is due to the other Rails' generators
expecting to have a "model" type class available, but Rails is unable to
load it.  Just adding the classes does not work as the generated
controllers, form helpers, rendering process, etc. are so tightly
coupled to the `ActiveModel` API interface that all sorts of things
start failing.