Fix failing builds on Rails 4.2.0-beta
Rails updated the way route generation errors are raised. Specifically,
this applies only to internal server code which generates a route and
not routes passed in from a request. This change comes from
rails/rails#16229. In a nutshell, the change is
`ActionController::UrlGenerationError` no longer inherits from
`ActionController::RoutingError`.

This updates the cukes which expect the unknown routes used in the
internal rails controller code to raise the correct base error
`UrlGenerationError`.

The current routing matchers do not require updating as they do not rely
on the internal route generation code. Instead they are utilizing the
routing engine's path for parsing routes passed from requests. This
path still raises a `ActionController::RoutingError`.