Fix failing scenario caused by a behavior of Rail.root with symlink
The failing scenario was caused by a change in Rails that affects
Rails.root when a Rails::Application is defined in symlinked file.

https://github.com/rails/rails/pull/17782

We've been set up Aruba workspace in each Cucumber scenario with
symlinks, but this tends to behave differently from real files.
Since the example_app project is not so large, we can simply clone it in
each scenario.