Reset Capybara's monkey patch on old versions.
Apps just upgrading to RSpec 3 are more likely to have several
conditions which may cause a poor user experience:

- Version of Capybara prior to 2.4.0
- Loading Capybara before rspec-rails in `spec/spec_helper.rb`
- Not enabling monkey patch mode or have set:

  `config.expose_dsl_globally = false`

When these conditions are all met our `feature` alias is exposed
globally overwriting the Capybara monkey patch. This will cause all
feature specs to be skipped.

For a user attempting to upgrade an app this can be a very frustrating,
and confusing, "bug".