Fix load order :bug:
This fixes an "undefined method `fixture_path`" error caused in
`RSpec::Core::Configuration`. The issue occurs when a spec is created
before rspec-rails is loaded. The cause is rspec-core attempts to
include the `RSpec::Rails::FixtureSupport` module into the existing spec
as soon as we tell the configuration about it, but have not told the
configuration about the new related options.

This moves the including of the module until after we have told the
configuration about the dependent options.

This includes a spec which is named in a manner to attempt to ensure it
is the first spec loaded by rspec-core; which loads spec files
alphabetically. To ensure we test the issue the run script is updated to
also run this spec file standalone.

When discussing rspec/rspec-rails#1355 it was agreed that we need to
consider removing this global inclusion of the fixture support. Instead
favoring including it in the `RailsExampleGroup`. A deprecation comment
has been included to remind us later.