Fix regression issue with `preview_path=`
The `preview_path` feature of `ActionMailer` was added in Rails 4.1.
Versions before that release are not able to handle that configuration.
The regression issue occurred in part to the spec suite overlooking
changes to the environment and how that affects the shelling out.

Additionally, the previous changes noted that the configuration for a
particular module always responds to a setting; even when it should not.
However, while this logic was noted for `show_previews`, a few lines
further down, that same logic was overlooked for `preview_path`.

It appears the only available methods for detecting `ActionMailer`, and
it's features, is tasting the top-level `config` in addition to checking
the Rails version string.