Ensure on Rails 5 and above we don't use `render :text`.
`render :text` has been deprecated for a while now, and is removed in
rails 5.1. It has been replaced by either `render :plain` or `render
:html` (depending on which `Content-Encoding` header you want). This
patch uses our existing cucumber tags `@rails_pre_5` and `@rails_post_5`
to switch which we use depending on the Rails verison. We can't simply
replace all the references here because Rails 3 does not have this
option.