Update request spec scaffold template
Currently, when scaffolds are generated inside a Rails Engine, it generates requests specs that fail without modification. This happens because the Engine routes are not mounted under Rails.application.

To fix this, Engine routes have to be included so that url and path helpers can be usable from within the spec file.